refactor authloading
