+++ src/router/index.js
@@ -1,15 +1,15 @@
 import React from "react";
-import { View, Text, Button, ActivityIndicator, StatusBar } from "react-native";
+import { View, Text } from "react-native";
 import MdIcons from 'react-native-vector-icons/MaterialCommunityIcons';
 import { createAppContainer, createSwitchNavigator} from 'react-navigation';
 import { createStackNavigator } from 'react-navigation-stack';
 import { createBottomTabNavigator } from 'react-navigation-tabs';
-import AsyncStorage from '@react-native-community/async-storage';
 
 // screen
 import LoginScreen from 'smartchef/src/scenes/login/login.container';
 import RegisterScreen from 'smartchef/src/scenes/register/register.container';
 import MainScreen from 'smartchef/src/components/MainScreen';
+import AuthLoadingScreen from 'smartchef/src/scenes/authLoading/authLoading.screen.container';
 // import ChatScreen from 'smartchef/src//components/ChatScreen';
 import DetailScreen from 'smartchef/src/components/DetailScreen';
 // import HomeScreen from 'smartchef/src/scenes/home/home.container';
@@ -17,33 +17,6 @@ import DetailScreen from 'smartchef/src/components/DetailScreen';
 // import AccountScreen from 'smartchef/src/scenes/account/account.screen';
 // utils
 import { Colors } from 'smartchef/src/styles/Colors';
-
-class AuthLoadingScreen extends React.Component {
-  constructor(props) {
-    super(props);
-    this._bootstrapAsync();
-  }
-
-  // Fetch the token from storage then navigate to our appropriate place
-  _bootstrapAsync = async () => {
-    const userToken = await AsyncStorage.getItem('@smartchefUser');
-
-    // This will switch to the App screen or Auth screen and this loading
-    // screen will be unmounted and thrown away.
-    console.log("usertoken", userToken);
-    this.props.navigation.navigate(userToken ? 'AppMain' : 'SignIn');
-  };
-
-  // Render any loading content that you like here
-  render() {
-    return (
-      <View>
-        <ActivityIndicator />
-        <StatusBar barStyle="default" backgroundColor="orange" />
-      </View>
-    );
-  }
-}
 
 const defaultScreen = () => {
   return (
